requirements.txt
aiofiles>=23.0.0
anthropic>=0.40.0
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
sqlalchemy[asyncio]>=2.0.0
greenlet>=3.0.0
asyncpg>=0.29.0
psycopg2-binary>=2.9.9
networkx>=3.3
scipy>=1.13.0
pandas>=2.2.0
numpy>=1.26.0
ortools>=9.9.3963
pydantic>=2.6.0
pydantic-settings>=2.2.0
python-dotenv>=1.0.0
httpx>=0.27.0
